refactor: 옵션 로직 분리
